+++ core/src/com/unciv/MainMenuScreen.kt
@@ -18,6 +18,7 @@ import com.unciv.models.ruleset.RulesetCache
 import com.unciv.ui.multiplayer.MultiplayerScreen
 import com.unciv.ui.mapeditor.*
 import com.unciv.models.metadata.GameSetupInfo
+import com.unciv.models.ruleset.Ruleset
 import com.unciv.ui.civilopedia.CivilopediaScreen
 import com.unciv.ui.crashhandling.launchCrashHandling
 import com.unciv.ui.crashhandling.postCrashHandlingRunnable
@@ -28,6 +29,7 @@ import com.unciv.ui.popup.*
 import com.unciv.ui.saves.LoadGameScreen
 import com.unciv.ui.utils.*
 import com.unciv.ui.utils.UncivTooltip.Companion.addTooltip
+import com.unciv.ui.worldscreen.mainmenu.WorldScreenMenuPopup
 
 class MainMenuScreen: BaseScreen() {
     private val backgroundTable = Table().apply { background= ImageGetter.getBackground(Color.WHITE) }
@@ -98,7 +100,7 @@ class MainMenuScreen: BaseScreen() {
 
         if (game.gameSaver.autosaveExists()) {
             val resumeTable = getMenuButton("Resume","OtherIcons/Resume", 'r')
-                { autoLoadGame() }
+                { resumeGame() }
             column1.add(resumeTable).row()
         }
 
@@ -165,7 +167,14 @@ class MainMenuScreen: BaseScreen() {
     }
 
 
-    private fun autoLoadGame() {
+    private fun resumeGame() {
+        val curWorldScreen = game.getWorldScreenOrNull()
+        if (curWorldScreen != null) {
+            game.resetToWorldScreen()
+            curWorldScreen.popups.filterIsInstance(WorldScreenMenuPopup::class.java).forEach(Popup::close)
+            return
+        }
+
         val loadingPopup = Popup(this)
         loadingPopup.addGoodSizedLabel("Loading...")
         loadingPopup.open()
@@ -192,12 +201,19 @@ class MainMenuScreen: BaseScreen() {
                 return@launchCrashHandling
             }
 
-            postCrashHandlingRunnable { /// ... and load it into the screen on main thread for GL context
+            if (savedGame.gameParameters.isOnlineMultiplayer) {
                 try {
-                    game.loadGame(savedGame)
-                    dispose()
+                    game.onlineMultiplayer.loadGame(savedGame)
                 } catch (oom: OutOfMemoryError) {
                     outOfMemory()
+                }
+            } else {
+                postCrashHandlingRunnable { /// ... and load it into the screen on main thread for GL context
+                    try {
+                        game.loadGame(savedGame)
+                    } catch (oom: OutOfMemoryError) {
+                        outOfMemory()
+                    }
                 }
             }
         }
@@ -230,9 +246,13 @@ class MainMenuScreen: BaseScreen() {
     }
 
     private fun openCivilopedia() {
-        val ruleset =RulesetCache[game.settings.lastGameSetup?.gameParameters?.baseRuleset]
-            ?: RulesetCache[BaseRuleset.Civ_V_GnK.fullName]
-            ?: return
+        val rulesetParameters = game.settings.lastGameSetup?.gameParameters
+        val ruleset = if (rulesetParameters == null)
+                RulesetCache[BaseRuleset.Civ_V_GnK.fullName] ?: return
+                else RulesetCache.getComplexRuleset(rulesetParameters)
+        UncivGame.Current.translations.translationActiveMods = ruleset.mods
+        ImageGetter.setNewRuleset(ruleset)
+        setSkin()
         game.setScreen(CivilopediaScreen(ruleset, this))
     }