+++ core/src/com/unciv/models/metadata/GameSettings.kt
@@ -6,6 +6,7 @@ import com.unciv.Constants
 import com.unciv.UncivGame
 import com.unciv.ui.utils.Fonts
 import java.text.Collator
+import java.time.Duration
 import java.util.*
 import kotlin.collections.HashSet
 
@@ -43,19 +44,14 @@ class GameSettings {
     var showPixelUnits: Boolean = true
     var showPixelImprovements: Boolean = true
     var continuousRendering = false
-    var userId = ""
-    var multiplayerTurnCheckerEnabled = true
-    var multiplayerTurnCheckerPersistentNotificationEnabled = true
-    var multiplayerTurnCheckerDelayInMinutes = 5
     var orderTradeOffersByAmount = true
+    var confirmNextTurn = false
     var windowState = WindowState()
     var isFreshlyCreated = false
     var visualMods = HashSet<String>()
     var useDemographics: Boolean = false
 
-
-    var multiplayerServer = Constants.dropboxMultiplayerServer
-
+    var multiplayer = GameSettingsMultiplayer()
 
     var showExperimentalWorldWrap = false // We're keeping this as a config due to ANR problems on Android phones for people who don't know what they're doing :/
 
@@ -71,10 +67,13 @@ class GameSettings {
     /** Maximum zoom-out of the map - performance heavy */
     var maxWorldZoomOut = 2f
 
+    /** used to migrate from older versions of the settings */
+    var version: Int? = null
+
     init {
         // 26 = Android Oreo. Versions below may display permanent icon in notification bar.
         if (Gdx.app?.type == Application.ApplicationType.Android && Gdx.app.version < 26) {
-            multiplayerTurnCheckerPersistentNotificationEnabled = false
+            multiplayer.turnCheckerPersistentNotificationEnabled = false
         }
     }
 
@@ -154,4 +153,15 @@ enum class LocaleCode(var language: String, var country: String) {
     Turkish("tr", "TR"),
     Ukrainian("uk", "UA"),
     Vietnamese("vi", "VN"),
+}
+
+class GameSettingsMultiplayer {
+    var userId = ""
+    var server = Constants.dropboxMultiplayerServer
+    var turnCheckerEnabled = true
+    var turnCheckerPersistentNotificationEnabled = true
+    var turnCheckerDelay = Duration.ofMinutes(5)
+    var statusButtonInSinglePlayer = false
+    var currentGameRefreshDelay = Duration.ofSeconds(10)
+    var allGameRefreshDelay = Duration.ofMinutes(5)
 }